+++ app/build.gradle
@@ -21,7 +21,12 @@ android {
 dependencies {
     implementation fileTree(include: ['*.jar'], dir: 'libs')
     implementation 'com.android.support:appcompat-v7:27.1.1'
+<<<<<<< HEAD
     implementation 'com.android.support.constraint:constraint-layout:1.1.0'
+=======
+    implementation 'com.android.support:exifinterface:27.1.1'
+    implementation 'com.android.support.constraint:constraint-layout:1.1.3'
+>>>>>>> first issue solving
     implementation 'com.android.support:recyclerview-v7:27.1.1'
     implementation 'com.android.support:design:27.1.1'
     implementation 'com.android.support:support-v4:27.1.1'
