+++ app/src/main/java/africa/apeiron/batafind/fragments/GROUPED_CATEGORY/Profile_Decorator.java
@@ -1,16 +1,13 @@
-package africa.apeiron.batafind.FRAGMENTS.GROUPED_CATEGORY;
+package africa.apeiron.batafind.fragments.GROUPED_CATEGORY;
 
 import android.graphics.Rect;
 import android.support.v7.widget.RecyclerView;
 import android.view.View;
 
-/**
- * Created by Granson on 25-Mar-18.
- */
-public class GCSpacesItemDecoration extends RecyclerView.ItemDecoration {
+public class Profile_Decorator extends RecyclerView.ItemDecoration {
     private final int mSpace;
 
-    public GCSpacesItemDecoration(int space) {
+    public Profile_Decorator(int space) {
         this.mSpace = space;
     }
