+++ app/src/main/java/africa/apeiron/batafind/item_display/DisplayProducts.java
@@ -1,11 +1,10 @@
-package africa.apeiron.batafind.ITEM_DISPLAY;
+package africa.apeiron.batafind.item_display;
 
 import android.content.Intent;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.support.v7.widget.RecyclerView;
 import android.support.v7.widget.StaggeredGridLayoutManager;
-import android.support.v7.widget.Toolbar;
 
 import java.util.ArrayList;
