+++ app/src/main/java/africa/apeiron/batafind/profile/Personal_Account.java
@@ -1,12 +1,9 @@
-package africa.apeiron.batafind.PROFILE;
+package africa.apeiron.batafind.profile;
 
-import android.content.Context;
 import android.os.Bundle;
 import android.support.design.widget.AppBarLayout;
 import android.support.design.widget.CollapsingToolbarLayout;
 import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentManager;
-import android.support.v4.app.FragmentStatePagerAdapter;
 import android.support.v4.view.ViewPager;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.Toolbar;
@@ -16,10 +13,10 @@ import com.ogaclejapan.smarttablayout.SmartTabLayout;
 import com.ogaclejapan.smarttablayout.utils.v4.FragmentPagerItemAdapter;
 import com.ogaclejapan.smarttablayout.utils.v4.FragmentPagerItems;
 
-import africa.apeiron.batafind.FRAGMENTS.FragmentHome;
-import africa.apeiron.batafind.FRAGMENTS.FragmentKids;
-import africa.apeiron.batafind.FRAGMENTS.FragmentLadies;
-import africa.apeiron.batafind.FRAGMENTS.FragmentMen;
+import africa.apeiron.batafind.fragments.FragmentHome;
+import africa.apeiron.batafind.fragments.FragmentKids;
+import africa.apeiron.batafind.fragments.FragmentLadies;
+import africa.apeiron.batafind.fragments.FragmentMen;
 import africa.apeiron.batafind.R;
