+++ app/src/main/res/layout/activity_personal_account.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context="africa.apeiron.batafind.PROFILE.Personal_Account">
+    tools:context="africa.apeiron.batafind.profile.Personal_Account">
 
     <android.support.design.widget.AppBarLayout
         android:id="@+id/app_bar"
