+++ app/src/main/res/layout/activity_profile.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context="africa.apeiron.batafind.PROFILE.ProfileActivity">
+    tools:context="africa.apeiron.batafind.profile.ProfileActivity">
 
     <android.support.design.widget.CoordinatorLayout
         android:id="@+id/coordinatorLayout"
