+++ app/src/main/res/layout/content_personal_account.xml
@@ -7,14 +7,12 @@
     android:fillViewport="true"
     android:layout_gravity="fill_vertical"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="africa.apeiron.batafind.PROFILE.Personal_Account"
+    tools:context="africa.apeiron.batafind.profile.Personal_Account"
     tools:showIn="@layout/activity_personal_account">
 
     <RelativeLayout
-        xmlns:android="http://schemas.android.com/apk/res/android"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        xmlns:app="http://schemas.android.com/apk/res-auto"
         android:background="@android:color/transparent">
 
             <com.ogaclejapan.smarttablayout.SmartTabLayout
