+++ app/src/main/res/layout/home_activity.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:background="#fff"
-    tools:context="africa.apeiron.batafind.HOME.HomeActivity">
+    tools:context="africa.apeiron.batafind.home.HomeActivity">
 
     <android.support.v4.view.ViewPager
         android:id="@+id/pager"
