+++ app/src/main/res/layout/new_arrival_item.xml
@@ -10,8 +10,7 @@
     android:elevation="10dp"
     app:cardUseCompatPadding="true">
 
-    <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-        xmlns:app="http://schemas.android.com/apk/res-auto"
+    <RelativeLayout
         android:layout_width="150dp"
         android:layout_height="150dp"
         android:background="@drawable/round_back">
@@ -37,7 +36,7 @@
             android:paddingEnd="5dp"
             android:paddingStart="5dp"
             android:text="Item Name"
-            android:textSize="10sp"
+            android:textSize="12sp"
             android:textColor="#353535"/>
 
         <TextView
@@ -53,7 +52,7 @@
             android:text="9-11-2017"
             android:textAlignment="viewEnd"
             android:textColor="#353535"
-            android:textSize="10sp" />
+            android:textSize="12sp" />
 
     </RelativeLayout>
