+++ app/src/main/res/menu/menu_personal__account.xml
@@ -1,7 +1,7 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="africa.apeiron.batafind.PROFILE.Personal_Account">
+    tools:context="africa.apeiron.batafind.profile.Personal_Account">
     <item
         android:id="@+id/action_settings"
         android:orderInCategory="100"
