+++ app/src/main/res/values/strings.xml
@@ -13,7 +13,7 @@
     <string name="todo">TODO</string>
     <string name="NA"><ul>New arrivals</ul></string>
     <string name="PD"><ul>Popular products</ul></string>
-    <string name="more">more...</string>
+    <string name="more">more</string>
     <string name="title_activity_personal__account">Personal_Account</string>
     <string name="large_text">
         "Material is the metaphor.\n\n"
